+++ openweather/model.go
@@ -12,8 +12,8 @@ type WeatherData struct {
 }
 
 type CoordsData struct {
-	Longitude float32 `json:"lon"`
-	Latitude  float32 `json:"lat"`
+	Longitude float64 `json:"lon"`
+	Latitude  float64 `json:"lat"`
 }
 
 type ReadableWeather struct {
@@ -24,21 +24,21 @@ type ReadableWeather struct {
 }
 
 type EnvironmentalData struct {
-	Temperature         float32 `json:"temp"`
-	Pressure            float32 `json:"pressure"`
+	Temperature         float64 `json:"temp"`
+	Pressure            float64 `json:"pressure"`
 	Humidity            int     `json:"humidity"`
-	MinTemperature      float32 `json:"temp_min"`
-	MaxTemperature      float32 `json:"temp_max"`
-	SeaLevelPressure    float32 `json:"sea_level"`
-	GroundLevelPressure float32 `json:"grnd_level"`
+	MinTemperature      float64 `json:"temp_min"`
+	MaxTemperature      float64 `json:"temp_max"`
+	SeaLevelPressure    float64 `json:"sea_level"`
+	GroundLevelPressure float64 `json:"grnd_level"`
 }
 
 type WindData struct {
-	Speed     float32 `json:"speed"`
-	Direction float32 `json:"deg"`
+	Speed     float64 `json:"speed"`
+	Direction float64 `json:"deg"`
 }
 
 type Precipitation struct {
-	OneHour   int `json:"1h"`
-	ThreeHour int `json:"3h"`
+	OneHour   float64 `json:"1h"`
+	ThreeHour float64 `json:"3h"`
 }
